+++ packages/account-api/src/api/index.ts
@@ -1,2 +1,8 @@
+export * from './bip44';
 export * from './group';
 export * from './wallet';
+export type * from './provider';
+export * from './selector';
+export * from './multichain';
+
+// NOTE: Do not export the `internal` sub-path.