+++ packages/account-api/src/api/internal/index.ts
@@ -1,0 +1,3 @@
+// This folder is meant to be used internally and should not be exported.
+
+export * from './utils';