+++ packages/keyring-utils/src/superstruct.ts
@@ -1,4 +1,9 @@
-import { Struct, assert, object as stObject } from '@metamask/superstruct';
+import {
+  Struct,
+  assert,
+  object as stObject,
+  type as stType,
+} from '@metamask/superstruct';
 import type {
   Infer,
   Context,
@@ -55,8 +60,8 @@ export type ObjectType<Schema extends ObjectSchema> = Simplify<
 >;
 
 /**
- * Change the return type of a superstruct object struct to support exact
- * optional properties.
+ * Change the return type of a superstruct's `object` function to support
+ * exact optional properties.
  *
  * @param schema - The object schema.
  * @returns A struct representing an object with a known set of properties.
@@ -65,6 +70,20 @@ export function object<Schema extends ObjectSchema>(
   schema: Schema,
 ): Struct<ObjectType<Schema>, Schema> {
   return stObject(schema) as any;
+}
+
+/**
+ * Change the return type of a superstruct's `type` function to support
+ * exact optional properties.
+ *
+ * @param schema - The object schema.
+ * @returns A struct representing an object with a known set of properties
+ * and ignore unknown properties.
+ */
+export function type<Schema extends ObjectSchema>(
+  schema: Schema,
+): Struct<ObjectType<Schema>, Schema> {
+  return stType(schema) as any;
 }
 
 /**