+++ packages/keyring-snap-bridge/CHANGELOG.md
@@ -7,6 +7,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [16.0.0]
+
+### Changed
+
+- **BREAKING:** Bump `@metamask/keyring-api` from `^19.0.0` to `^20.0.0` ([#347](https://github.com/MetaMask/accounts/pull/347))
+  - This is similar to version `15.0.0`, but since the `@metamask/keyring-api` was not versioned as a new major, we have to re-bump it here again.
+- **BREAKING:** Bump `@metamask/keyring-internal-api` from `^7.0.0` to `^8.0.0` ([#347](https://github.com/MetaMask/accounts/pull/347))
+  - This is similar to version `15.0.0`, but since the `@metamask/keyring-internal-api` was not versioned as a new major, we have to re-bump it here again.
+
+## [15.0.0]
+
+### Changed
+
+- **BREAKING:** Add `isAnyAccountTypeAllowed` flag to `SnapKeyring` constructor ([#322](https://github.com/MetaMask/accounts/pull/322))
+  - The `SnapKeyring` constructor now accepts an options object instead of individual parameters.
+  - The new `isAnyAccountTypeAllowed` flag defaults to `false`.
+  - This flag only affects account creation and update; account deletion is unaffected by it.
+
+## [14.0.0]
+
+### Changed
+
+- **BREAKING:** Bump `@metamask/keyring-api` from `^18.0.0` to `^19.0.0` ([#317](https://github.com/MetaMask/accounts/pull/317))
+  - The `KeyringAccount.options` field is now partially typed.
+- **BREAKING:** Bump `@metamask/keyring-internal-api` from `^6.2.0` to `^7.0.0` ([#317](https://github.com/MetaMask/accounts/pull/317))
+  - The `InternalAccount.options` field is now partially typed.
+
 ## [13.0.0]
 
 ### Changed
@@ -522,7 +549,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/MetaMask/accounts/compare/@metamask/eth-snap-keyring@13.0.0...HEAD
+[Unreleased]: https://github.com/MetaMask/accounts/compare/@metamask/eth-snap-keyring@16.0.0...HEAD
+[16.0.0]: https://github.com/MetaMask/accounts/compare/@metamask/eth-snap-keyring@15.0.0...@metamask/eth-snap-keyring@16.0.0
+[15.0.0]: https://github.com/MetaMask/accounts/compare/@metamask/eth-snap-keyring@14.0.0...@metamask/eth-snap-keyring@15.0.0
+[14.0.0]: https://github.com/MetaMask/accounts/compare/@metamask/eth-snap-keyring@13.0.0...@metamask/eth-snap-keyring@14.0.0
 [13.0.0]: https://github.com/MetaMask/accounts/compare/@metamask/eth-snap-keyring@12.2.0...@metamask/eth-snap-keyring@13.0.0
 [12.2.0]: https://github.com/MetaMask/accounts/compare/@metamask/eth-snap-keyring@12.1.1...@metamask/eth-snap-keyring@12.2.0
 [12.1.1]: https://github.com/MetaMask/accounts/compare/@metamask/eth-snap-keyring@12.1.0...@metamask/eth-snap-keyring@12.1.1