+++ packages/keyring-snap-sdk/CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.0.0]
+
+### Changed
+
+- **BREAKING:** Bump `@metamask/keyring-api` from `^19.0.0` to `^20.0.0` ([#347](https://github.com/MetaMask/accounts/pull/347))
+  - Add generic account type.
+
+## [5.0.0]
+
+### Changed
+
+- **BREAKING:** Bump `@metamask/keyring-api` from `^18.0.0` to `^19.0.0` ([#317](https://github.com/MetaMask/accounts/pull/317))
+  - The `KeyringAccount.options` field is now partially typed.
+- Bump `@metamask/snaps-sdk` dependency from `^6.16.0` to `^9.0.0` ([#273](https://github.com/MetaMask/accounts/pull/273)), ([#304](https://github.com/MetaMask/accounts/pull/304))
+
 ## [4.0.0]
 
 ### Changed
@@ -87,7 +102,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - This new version fixes a bug with CJS re-exports.
 - Initial release ([#24](https://github.com/MetaMask/accounts/pull/24))
 
-[Unreleased]: https://github.com/MetaMask/accounts/compare/@metamask/keyring-snap-sdk@4.0.0...HEAD
+[Unreleased]: https://github.com/MetaMask/accounts/compare/@metamask/keyring-snap-sdk@6.0.0...HEAD
+[6.0.0]: https://github.com/MetaMask/accounts/compare/@metamask/keyring-snap-sdk@5.0.0...@metamask/keyring-snap-sdk@6.0.0
+[5.0.0]: https://github.com/MetaMask/accounts/compare/@metamask/keyring-snap-sdk@4.0.0...@metamask/keyring-snap-sdk@5.0.0
 [4.0.0]: https://github.com/MetaMask/accounts/compare/@metamask/keyring-snap-sdk@3.2.0...@metamask/keyring-snap-sdk@4.0.0
 [3.2.0]: https://github.com/MetaMask/accounts/compare/@metamask/keyring-snap-sdk@3.1.0...@metamask/keyring-snap-sdk@3.2.0
 [3.1.0]: https://github.com/MetaMask/accounts/compare/@metamask/keyring-snap-sdk@3.0.1...@metamask/keyring-snap-sdk@3.1.0